removal of auth-management service
