--- packages/client-core/src/user/components/Auth/VerifyEmail.tsx
@@ -1,62 +1,0 @@
-/*
-CPAL-1.0 License
-
-The contents of this file are subject to the Common Public Attribution License
-Version 1.0. (the "License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-https://github.com/EtherealEngine/etherealengine/blob/dev/LICENSE.
-The License is based on the Mozilla Public License Version 1.1, but Sections 14
-and 15 have been added to cover use of software over a computer network and 
-provide for limited attribution for the Original Developer. In addition, 
-Exhibit A has been modified to be consistent with Exhibit B.
-
-Software distributed under the License is distributed on an "AS IS" basis,
-WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the
-specific language governing rights and limitations under the License.
-
-The Original Code is Ethereal Engine.
-
-The Original Developer is the Initial Developer. The Initial Developer of the
-Original Code is the Ethereal Engine team.
-
-All portions of the code written by the Ethereal Engine team are Copyright © 2021-2023 
-Ethereal Engine. All Rights Reserved.
-*/
-
-import React, { useEffect } from 'react'
-import { useTranslation } from 'react-i18next'
-
-import Box from '@etherealengine/ui/src/primitives/mui/Box'
-import Container from '@etherealengine/ui/src/primitives/mui/Container'
-import Typography from '@etherealengine/ui/src/primitives/mui/Typography'
-
-import { AuthService } from '../../services/AuthService'
-import styles from './index.module.scss'
-
-interface Props {
-  token: string
-}
-
-export const VerifyEmail = ({ token }: Props): JSX.Element => {
-  const { t } = useTranslation()
-
-  useEffect(() => {
-    AuthService.verifyEmail(token)
-  }, [])
-
-  return (
-    <Container component="main" maxWidth="md">
-      <div className={styles.paper}>
-        <Typography component="h1" variant="h5">
-          {t('user:auth.verifyEmail.header')}
-        </Typography>
-
-        <Box mt={3}>
-          <Typography variant="body2" color="textSecondary" align="center">
-            {t('user:auth.verifyEmail.processing')}
-          </Typography>
-        </Box>
-      </div>
-    </Container>
-  )
-}
