+++ packages/server-core/src/hooks/send-invite.ts
@@ -42,9 +42,20 @@ import { Application } from '../../declarations'
 import logger from '../ServerLogger'
 import config from '../appconfig'
 import { InviteParams } from '../social/invite/invite.class'
-import { getInviteLink, sendEmail, sendSms } from '../user/auth-management/auth-management.utils'
 
 const emailAccountTemplatesPath = path.join(appRootPath.path, 'packages', 'server-core', 'email-templates', 'invite')
+
+/**
+ * A method which get an invite link
+ *
+ * @param type
+ * @param id of accept invite
+ * @param passcode
+ * @returns invite link
+ */
+export function getInviteLink(type: string, id: string, passcode: string): string {
+  return `${config.server.url}/a-i/${id}?t=${passcode}`
+}
 
 async function generateEmail(
   app: Application,
@@ -91,7 +102,8 @@ async function generateEmail(
     html: compiledHTML
   }
 
-  return await sendEmail(app, email)
+  email.html = email.html.replace(/&amp;/g, '&')
+  await app.service('email').create(email)
 }
 
 async function generateSMS(
@@ -134,7 +146,12 @@ async function generateSMS(
     mobile,
     text: compiledHTML
   }
-  return await sendSms(app, sms)
+
+  await app
+    .service('sms')
+    .create(sms, null!)
+    .then(() => logger.info('Sent SMS'))
+    .catch((err: any) => logger.error(err, `Error sending SMS: ${err.message}`))
 }
 
 // This will attach the owner ID in the contact while creating/updating list item
